feat: Rename AccessDeniedResponse to ForbiddenResponse
This commit renames the `AccessDeniedResponse` class to `ForbiddenResponse`. This change improves the clarity and accuracy of the class name, as it more accurately reflects the HTTP status code (403 Forbidden) that the response represents. The class is used to represent API access denied scenarios and is part of the swagger documentation.